refactor: move extensions/config/validators/cli/middleware into app/core
Consolidates the cross-cutting "project layer" (Django settings.py/wsgi.py
equivalent) into app/core/: extensions.py (shared db/login/mail/cache/
limiter singletons), config.py, validators/env.py, cli.py, and the
middleware/ package. Pure git-mv plus import-path updates across ~40
call sites; no behavior change. Verified: create_app() builds cleanly
with all 224 routes intact, and the pytest suite is unchanged from
baseline (25 pre-existing failures unrelated to this move, 99 passing).
