+++ app/__init__.py
@@ -5,9 +5,9 @@ from typing import Optional, Dict
 from dotenv import load_dotenv
 from flask import Flask, jsonify, request, render_template
 
-from .extensions import db, login_manager, mail, migrate, cache, limiter
-from .config import config_by_name
-from .validators.env import run_all_validations
+from .core.extensions import db, login_manager, mail, migrate, cache, limiter
+from .core.config import config_by_name
+from .core.validators.env import run_all_validations
 
 
 def create_app(test_config: Optional[Dict] = None) -> Flask:
@@ -218,7 +218,7 @@ def create_app(test_config: Optional[Dict] = None) -> Flask:
         return {'layout': 'layouts/_fragment.html' if is_hx else 'base.html'}
 
     # Register middleware
-    from .middleware import register_all_middleware
+    from .core.middleware import register_all_middleware
     register_all_middleware(app)
 
     # Register blueprints
@@ -233,7 +233,7 @@ def create_app(test_config: Optional[Dict] = None) -> Flask:
     app.register_blueprint(metrics_bp)
 
     # Register CLI commands
-    from .cli import register_cli_commands
+    from .core.cli import register_cli_commands
     register_cli_commands(app)
 
     # Global Error Handlers
