+++ app/blueprints/auth.py
@@ -3,7 +3,7 @@ from flask import Blueprint, request, jsonify, render_template, redirect
 from flask_login import login_user, logout_user, current_user, login_required
 from app.models import User, UserRole, SchoolStatus, PasswordResetToken, School
 from app.helpers import is_valid_email, log_activity, validate_password
-from app.extensions import limiter, db
+from app.core.extensions import limiter, db
 from app.services.auth_service import (
     register_school, verify_email_token,
     request_password_reset, reset_password, register_user,
