+++ app/models/__init__.py
@@ -1,4 +1,4 @@
-from app.extensions import db
+from app.core.extensions import db
 
 from .school import School, SchoolStatus
 from .user import User, UserRole
