+++ app/models/activity_log.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from typing import Optional
 from sqlalchemy.orm import Mapped, mapped_column, relationship
-from app.extensions import db
+from app.core.extensions import db
 from app.helpers import get_jakarta_now
