+++ app/models/school.py
@@ -2,7 +2,7 @@ import enum
 from datetime import datetime
 from typing import List, Optional
 from sqlalchemy.orm import Mapped, mapped_column, relationship
-from app.extensions import db
+from app.core.extensions import db
 from app.helpers import get_jakarta_now
