+++ app/models/whats_new.py
@@ -1,4 +1,4 @@
-from app.extensions import db
+from app.core.extensions import db
 from datetime import datetime
 from typing import Optional, List
 from sqlalchemy.orm import Mapped, mapped_column
