+++ app/services/auth_service.py
@@ -1,5 +1,5 @@
 import re
-from app.extensions import db
+from app.core.extensions import db
 from app.models import (
     User, UserRole, School, SchoolStatus,
     EmailVerificationToken, PasswordResetToken
