+++ app/services/ticket_service.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from app.extensions import db
+from app.core.extensions import db
 from app.models import Ticket, TicketStatus, TicketMessage, UserRole
 from app.helpers import get_jakarta_now
